Update and rename Default.java to Portuguese.java
First translation. Will edit and spellcheck later. Any help appreciated.

Commands also need changing to portuguese. For now all commands are in English.
